Move non-client specific log related code out of client/logging.
