+++ pkg/client/logging/initcontext.go
@@ -9,13 +9,12 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/telepresenceio/telepresence/v2/pkg/log"
-
 	"github.com/sirupsen/logrus"
 
 	"github.com/datawire/dlib/dlog"
 	"github.com/telepresenceio/telepresence/v2/pkg/client"
 	"github.com/telepresenceio/telepresence/v2/pkg/filelocation"
+	"github.com/telepresenceio/telepresence/v2/pkg/log"
 )
 
 // loggerForTest exposes internals to initcontext_test.go
@@ -31,9 +30,9 @@ func InitContext(ctx context.Context, name string) (context.Context, error) {
 	logger.ReportCaller = true
 
 	if IsTerminal(int(os.Stdout.Fd())) {
-		logger.Formatter = NewFormatter("15:04:05.0000")
+		logger.Formatter = log.NewFormatter("15:04:05.0000")
 	} else {
-		logger.Formatter = NewFormatter("2006/01/02 15:04:05.0000")
+		logger.Formatter = log.NewFormatter("2006/01/02 15:04:05.0000")
 		dir, err := filelocation.AppUserLogDir(ctx)
 		if err != nil {
 			return ctx, err
